+++ src/utils/dssim-comparisons.sh
@@ -1,0 +1,25 @@
+#!/bin/sh
+
+echo "Cat image"
+dssim "../../img/cat-1285634_640.png" "../../img/cat-1285634_640-crunch.png"
+
+# sun's rays image
+echo "\n\nSun image"
+dssim "../../img/suns-rays-478249_640.png" "../../img/suns-rays-478249_640-crunch.png"
+
+# prairie image
+echo "\n\nPrairie image"
+dssim "../../img/prairie-679014_640.png" "../../img/prairie-679014_640-crunch.png"
+
+# Robot image
+echo "\n\nRobot image"
+dssim "../../img/robot-1214536_640.png" "../../img/robot-1214536_640-crunch.png"
+
+# Color circule image
+echo "\n\nColor circle image"
+dssim "../../img/colors-157474_640.png" "../../img/colors-157474_640-crunch.png"
+
+# Flowers image
+echo "\n\nFlowers image"
+dssim "../../img/flowers-67839_640.png" "../../img/flowers-67839_640-crunch.png"
+